# openapi/v1.yaml
openapi: "3.0.3"
info:
  title: EdgeFabric Controller API
  description: |
    REST API for the EdgeFabric distributed edge networking platform.
    Provides multi-tenant management, user authentication (password + TOTP 2FA),
    API key management, fleet management (nodes, node groups, SSH keys),
    audit logging, and status dashboard.
  version: "1.1.0"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8443
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/node-groups:
    get:
      tags: [Node Groups]
      summary: List node groups
      description: |
        Returns node groups for a tenant. Non-superusers see only their tenant's groups.
        SuperUsers must provide a tenant_id query parameter.
      parameters:
        - name: tenant_id
          in: query
          description: Required for SuperUser callers to scope the list.
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated list of node groups
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeGroupListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Node Groups]
      summary: Create a node group
      description: Tenant is inferred from caller's claims for non-superusers.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                tenant_id:
                  type: string
                  format: uuid
                  description: Required for SuperUser. Inferred from token for tenant-scoped users.
                name:
                  type: string
                description:
                  type: string
      responses:
        "201":
          description: Node group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse_NodeGroup"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Node group already exists
  /api/v1/node-groups/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [Node Groups]
      summary: Get a node group by ID
      description: Enforces tenant isolation for non-superuser callers.
      responses:
        "200":
          description: Node group details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse_NodeGroup"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Node Groups]
      summary: Delete a node group
      description: Enforces tenant isolation. Removes the group and its memberships.
      responses:
        "204":
          description: Node group deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/node-groups/{id}/nodes/{nodeId}:
    parameters:
      - $ref: "#/components/parameters/ID"
      - name: nodeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Node Groups]
      summary: Add a node to a group
      description: Both the group and the node must exist.
      responses:
        "204":
          description: Node added to group
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Node already in group
    delete:
      tags: [Node Groups]
      summary: Remove a node from a group
      responses:
        "204":
          description: Node removed from group
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/ssh-keys:
    get:
      tags: [SSH Keys]
      summary: List SSH keys
      description: Returns all SSH keys. Private keys are never included in responses.
      parameters:
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated list of SSH keys
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SSHKeyListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [SSH Keys]
      summary: Create an SSH key
      description: |
        Only SuperUsers can create SSH keys. SSH keys are a global resource
        (not tenant-scoped). The private key is stored encrypted at rest
        and never returned in API responses.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, public_key, private_key]
              properties:
                name:
                  type: string
                public_key:
                  type: string
                private_key:
                  type: string
                  description: Must be pre-encrypted or will be encrypted at rest.
      responses:
        "201":
          description: SSH key created (private key stripped from response)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse_SSHKey"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: SSH key already exists
  /api/v1/ssh-keys/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [SSH Keys]
      summary: Get an SSH key by ID
      description: Private key is never included in the response.
      responses:
        "200":
          description: SSH key details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse_SSHKey"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [SSH Keys]
      summary: Delete an SSH key
      description: Only SuperUsers can delete SSH keys.
      responses:
        "204":
          description: SSH key deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Session token (from /auth/login) or API key (prefixed with ef_).
        Example: `Authorization: Bearer <token>` or `Authorization: Bearer ef_<key>`
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
  schemas:
...
    NodeGroup:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SSHKey:
      type: object
      description: SSH key (private key is never returned in API responses).
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        public_key:
          type: string
        fingerprint:
          type: string
        created_at:
          type: string
          format: date-time
        last_rotated_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...
    DataResponse_NodeGroup:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/NodeGroup"
    DataResponse_SSHKey:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/SSHKey"
...
    NodeGroupListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/NodeGroup"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
    SSHKeyListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SSHKey"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"